Add files via upload
Measure of how long it took to run the analysis was added. After the stock analysis for a given year is calculated a message box appears stating how long it took to calculate. 

This was done by using the Timer function. At the end of the routine (after Next i and before End Sub) the endTime was input and a message box that subtracted the end time from the start time.